feat: [mergeability-follow-up] - [Split provider streaming files for backend-size]
